+++ R/options.R
@@ -49,7 +49,7 @@ r_process_options_default <- function() {
     stdout = "|",
     stderr = "|",
     error = getOption("callr.error", "error"),
-    cmdargs = c("--no-site-file", "--no-environ", "--slave",
+    cmdargs = c("--no-site-file", "--slave",
       "--no-save", "--no-restore"),
     system_profile = FALSE,
     user_profile = FALSE,