+++ tests/testthat/test-eval.R
@@ -74,3 +74,55 @@ test_that("stdout and stderr in the same file", {
 
   expect_equal(readLines(tmp), paste0("hello", 1:3))
 })
+
+test_that("profiles are used as requested", {
+  do <- function(system, user) {
+    tmp1 <- tempfile()
+    tmp2 <- tempfile()
+    on.exit(unlink(c(tmp1, tmp2)), add = TRUE)
+    cat("Sys.setenv(FOO = 'bar')\n", file = tmp1)
+    cat("Sys.setenv(NAH = 'doh')\n", file = tmp2)
+    withr::with_envvar(list(R_PROFILE = tmp1, R_PROFILE_USER = tmp2),  {
+      res <- r(
+        function() c(Sys.getenv("FOO", ""), Sys.getenv("NAH", "")),
+        system_profile = system, user_profile = user)
+    })
+  }
+
+  ## None
+  res <- do(FALSE, FALSE)
+  expect_equal(res, c("", ""))
+
+  ## System
+  res <- do(TRUE, FALSE)
+  expect_equal(res, c("bar", ""))
+
+  ## User
+  res <- do(FALSE, TRUE)
+  expect_equal(res, c("", "doh"))
+
+  ## Both
+  res <- do(TRUE, TRUE)
+  expect_equal(res, c("bar", "doh"))
+})
+
+test_that(".Renviron is used, but lib path is set over it", {
+  dir.create(tmp <- tempfile())
+  on.exit(unlink(tmp, recursive = TRUE), add = TRUE)
+  withr::with_dir(tmp, {
+    ## Create .Renviron file
+    dir.create("not")
+    dir.create("yes")
+    cat("R_LIBS=\"", file.path(getwd(), "not"), "\"\n",
+        sep = "", file = ".Renviron")
+    cat("FOO=bar\n", file = ".Renviron", append = TRUE)
+
+    res <- r(
+      function() list(.libPaths(), Sys.getenv("FOO")),
+      libpath = file.path(getwd(), "yes")
+    )
+  })
+
+  expect_equal(basename(res[[1]][1]), "yes")
+  expect_equal(res[[2]], "bar")
+})